Fix styling on PhotoShow, AlbumShow. Added session creditinals to AlbumShow and CommentForms

+++ frontend/components/comment/comment_show_item.jsx
@@ -1,11 +1,18 @@
 import React from 'react';
 
-const CommentShowItem = ({comment, deleteComment}) => {
-  return(
-    <div className='comment-show-item'>
+const CommentShowItem = ({comment, deleteComment, currentUser}) => {
+  let hiddenItem = <div className="hor-spacer"></div>;
+  debugger;
+  if (parseInt(currentUser.id) === comment.user_id) {
+    hiddenItem = (
       <div onClick={()=>(deleteComment(comment.id))} >
         <i className="fas fa-times-circle"></i>
       </div>
+    );
+  }
+  return(
+    <div className='comment-show-item'>
+      {hiddenItem}
       <div className="commenter-info">
         <img className="photo-show-user-info" src={comment.user_img ? `${comment.user_img}` :"https://i.stack.imgur.com/IHLNO.jpg"}></img>
       </div>
